pyproject.toml
[tool.poetry]
name = "zelt"
version = "1.0.0"
description = "Zalando end-to-end load testing"
authors = [
    "Thibaut Le Page <thibaut.le.page@zalando.de>",
    "Brian Maher <brian.maher@zalando.de>",
    "Oliwia Zaremba <oliwia.zaremba@zalando.de>"
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
har-transformer = "^1.0"
kubernetes = "^8.0"
locustio = "^0.9.0"
boto3 = "^1.9"
greenlet = "^0.4.15"
tenacity = "^5.0"
docopt = "^0.6.2"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.6"
pytest = "^4.3"
pylint = "^2.3"
pytest-watch = "^4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
zelt = "main:cli"
